refactor(racing): 테스트 코드 리팩터링
- List.of를 Arrays.asList()로 수정
- UtilTest를 racing.utils 패키지로 이동
- 불필요한 import 제거
