+++ src/test/java/racing/domain/CarTest.java
@@ -2,14 +2,11 @@ package racing.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-//@TestInstance(TestInstance.Lifecycle.PER_CLASS)
 public class CarTest {
     private Car car;
