+++ src/test/java/racing/utils/UtilTest.java
@@ -1,8 +1,7 @@
-package racing;
+package racing.utils;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.RepeatedTest;
-import racing.utils.RandomNumberGenerator;
 
 import static org.assertj.core.api.Assertions.assertThat;
